drive_streamlit_incremental_upload/requirements.txt
streamlit==1.40.2
pandas==2.2.3
openpyxl==3.1.5
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
